Added protocol folders for each feature
