+++ BookWyrmTests/CriticReviewViewModelTests.swift
@@ -1,14 +1,15 @@
 //
-//  BookWyrmTests.swift
+//  CriticReviewViewModelTest.swift
 //  BookWyrmTests
 //
-//  Created by Zaheer Moola on 2019/02/20.
+//  Created by Zaheer Moola on 2019/03/07.
 //  Copyright © 2019 DVT. All rights reserved.
 //
 
 import XCTest
+@testable import BookWyrm
 
-class BookWyrmTests: XCTestCase {
+class CriticReviewViewModelTests: XCTestCase {
 
     override func setUp() {
         // Put setup code here. This method is called before the invocation of each test method in the class.
